Move the "mode" JavaScript setting into JSConfig
The "mode" config setting tells the JavaScript code whether it should do
a normal assignment launch (`"basic-lti-launch"` mode) or whether it
should show the UI for the teacher to configure the assignment
(selecting the document for the assignment) (`"content-item-selection"`
mode).

The mode is now changed to default to `"basic-lti-launch"` mode, and it
gets explicitly changed to `"content-item-selection"` in two specific
places.

Previously the default was `"content-item-selection"` (but this default
was implemented in JavaScript not Python) and it had to be explicitly
set to `"basic-lti-launch"` and then in one place it had to be
explicitly set back to `"content-item-selection"` again _after_ changing
it to `"basic-lti-launch"`.