Move is_canvas into LTILaunchResource
Move `BasicLTILaunchViews.is_launched_by_canvas()` to
`LTILaunchResource.is_canvas` (a property).

This is necessary because future `JSConfig` code is going to need to
know whether the LMS is Canvas, and `JSConfig` can't access a
`BasicLTILaunchViews` method.

One thing that both `BasicLTILaunchViews` and `JSConfig` have access to
is the `LTILaunchResource` context, so they can both read
`context.is_canvas`.

Whether or not the LMS is Canvas also seems fitting as part of the
"context".

This is also in line with keeping views thin in general. One of the
reasons for keeping code out of views is that view code isn't reusable.
Context code is usable by the context itself, context helpers like
`JSConfig`, views, services that receive the context, etc.