Change from Prospector to PyLint
Remove the Prospector meta-linter and change `make lint` to call pylint,
pydocstyle and pycodestyle directly instead.

Prospector wasn't serving us well. It doesn't work with PyLint's
multiprocessing support so it's a lot slower -- running PyLint on one
CPU. It also seems to mess with the linters it runs: for example it
seems to have been disabling a lot of useful PyLint rules that we didn't
tell it to disable, resulting in a lot of messages if you run PyLint
directly.

What this PR does:

* Get rid of Prospector and just run the underlying tools directly. This
  is faster and gives us better control of the tools.

* Add PyLint, pycodestyle and pydocstyle configs

* I've also removed `make docstrings` and `make checkdocstrings`. These
  were being used to validate the docstrings, but PyLint's `docparams`
  extension can do at least some of that validation for us: it checks
  for missing or incorrect `:param`'s and `:raise`'s etc in docstrings.
  Just use that instead.