Use a hardcoded userid to create groups
Use a hardcoded userid (`acct:lms@lms.hypothes.is` in production,
`acct:lms@qa-lms.hypothes.is` in QA) as the group's creator
(`X-Forwarded-User`) when creating groups, instead of using the h user
corresponding to the LMS course teacher.

This means we can now create groups in response to either teacher or
student launches. We no longer need to crash when a student launches an
assignment and the course's group doesn't exist yet.

This will crash if the `lms` user doesn't exist in h, though, and
nothing creates that user yet.