Automatically create the "lms" user if it doesn't exist
If creating a group fails because the lms user doesn't exist in h yet,
then create the lms user in h and then try again to create the group.

The way this works is that:

1. The LMS app always uses the `lms@[qa-]lms.hypothes.is` user as the
   `X-Forwarded-User` when creating groups. If that user doesn't exist
   yet in the DB then the create-group request will 404

2. If the create-group request 404s then the LMS app will send a
   create-user API request to h to create the `lms@[qa-]lms.hypothes.is`
   and then try the create-group request again

3. If either the create-user or the second create-group request fails
   the LMS app will show an error page

In practice this means that we don't need to create the lms user in dev,
qa or prod ourselves. The LMS app will try to use the lms user for the
first time, get an error once, create the lms user, and continue from
then on without errors.

If the lms user ever gets deleted somehow (e.g. dev DBs getting reset)
the LMS app will just recreate it automatically.

The logic here is specific to the create-group API, which 404s if the
lms user doesn't exist. Hence the logic is implemented in the
create_group() method.

I don't know what other h APIs would do if given an X-Forwarded-User
that doesn't exist, and whether it would be possible to tell from the
response whether it failed because the X-Forwarded-User doesn't exist or
whether it failed for another reason.

In practice we only use X-Forwarded-User when creating groups, so
implementing this as part of create_group() is fine.