Simplify upsert_course_group()
Now that we use a machine user as the group's creator when upserting
groups, it's no longer necessary to make sure that only teachers (and
not students) create groups or to make sure that the group's creator
doesn't change from one teacher to another over time. This means that
`upsert_course_group()` no longer needs to try to _update_ the group and
then only upsert if that fails and the user is a teacher. It can simply
upsert the group.

`HAPI.update_group()` is no longer used at all.