Fix creating assignments in non-Canvas LMS's
A recent commit changed a bunch of code in `LTILaunchResource` to read
from `request.parsed_params` not `request.params`. This means the params
being read must be added to the validation schema that the view uses.

We forgot to update the `configure_module_item()` view and its
`ConfigureModuleItemSchema`.

Fixes https://github.com/hypothesis/lms/issues/1626