Break LaunchVerifier's dependency on ApplicationInstanceGetter
This is necessary because I need to make a change to
ApplicationInstanceGetter that will make it indirectly dependent on
LaunchVerifier, and since LaunchVerifier already depends on
ApplicationInstanceGetter that would result in a circular dependency.
Avoid that by changing LaunchVerifier to no longer depend on
ApplicationInstanceGetter:

* Split models.ApplicationInstance.get() out of
  ApplicationInstanceGetter.  This is the method for doing the actual
  SQLAlchemy query for an application instance. Move this onto the model
  class so that both ApplicationInstanceGetter and LaunchVerifier can
  call it.

* Change ApplicationInstanceGetter to call
  models.ApplicationInstance.get() instead of doing the SQLAlchemy query
  itself

* Change LaunchVerifier to call models.ApplicationInstance.get() instead
  of using ApplicationInstanceGetter

So now LaunchVerifier no longer depends on ApplicationInstanceGetter.
Instead both services depend on models.ApplicationInstance directly.