Simplify LTIUser to just have a display_name
Simplify models.LTIUser to just have a display_name field instead of
separate given_name, family_name and full_name fields.

We never actually use the given_name, family_name and full_name. We only
ever use a display_name that we generate from them (which we pass to the
h API as the user's h display name).

So just generate the display_name once when the LTIUser object is first
created, and don't hold on to the original three separate name parts.

The generating of a display_name from the lis_person_name_given,
lis_person_name_family and lis_person_name_full LTI params now happens
when creating the LTIUser in the first place (in LaunchParamsAuthSchema)
rather than every time we create an HUser from the LTIUser.

This means that when an LTIUser is serialized and deserialized by
BearerTokenSchema and CanvasOAuthCallbackSchema only the display_name
ever gets serialized or deserialized any more, not the separate name
parts.

The LTIUser class can now be constructed by passing it a display name
directly if you already have one:

    user = LTIUser(display_name="...", **kwargs)

Or, if you have separate given, family and full name parts, there's a
factory method that will generate a display_name from them for you:

    user = LTIUser.from_names("given", "family", "full", **kwargs)