Skip Via 3 routing for Google Drive documents
Got a little fancy here and introduced a simple model class here
to help re-use some of the logic. The idea is to make the central
bit nice and declarative (if is_pdf do this).

The document object also paves the way for us to specify that
something is a PDF for Canvas docs (in the next PR)