Add requestGroups RPC call
`requestGroups` is a non-standard RPC handler method in that its value is pre-loaded before the client makes the request. This value is resolved through an api endpoint call (sync) on app load. The awaited result of that call is returned to the client via  RPC. e.g. `requestGroups`.

The pre-loaded result is done in BasicLtiLauchApp.js and has retry support if for any reason the auth token expires or is invalided.

BasicLtiLauchApp will also wait to render its own iframe until the sync api endpoint (which returns the groups) is resolved.

This feature is primary used to support canvas sections.