Simplify iter_columns()
Change iter_columns() to return a list of column name strings, which is
what's actually needed in each of the three places where this method is
called. I have a need to call the method in a fourth place in a future
commit, and that place also needs column name strings not column
objects.

And just rename it to columns().

Also inline the iter_properties() method, which is only used in one
place in columns().

This also simplifies the tests.