Sort students by last, then first name
The client attempts to discover a last name based off the displayName, but if it does not exist, then it sorts by first name.

Additionally if there are more separators, then every word past the first is assumed to be part of the last name. This may be a false assumption but its a best effort with only having a displayName.

Finally, the displayName is changed from the service layer to show the last then first. e.g. “{last}, {first}”