Show student's sections in SpeedGrader launches
Show the learner's section groups, not the teacher's, in the client for
SpeedGrader launches.

If the "learner" dict is present in the JSON that the sync API is called
with, then get the learner's Canvas user ID from the learner dict and
use it to retrieve the learner's sections from Canvas and return those
instead of returning the teacher's sections.

Note that we still also retrieve the teacher's sections and sync them to
h: this ensures that the teacher's h user exists, that all of the
section groups exist in h, and that the teacher is a member of the
groups. Otherwise the SpeedGrader launch could potentially fail /
misbehave if h didn't contain groups for the learner's sections, if the
teacher's h user wasn't a member of those groups, or if the teacher's h
user didn't exist yet.

This unfortunately means that we're making two separate requests to the
Canvas API for every one SpeedGrader sync API request: first to retrieve
the teacher's groups and then to retrieve the learners.