Add requestGroups RPC call (#1709)
`requestGroups` is a non-standard RPC handler method in that its value is pre-loaded before the client makes the request. This value is resolved through an api endpoint call (sync) on app load. The awaited result of that call is returned to the client via  RPC. e.g. `requestGroups`.

The pre-loaded result is done in BasicLtiLaunchApp.js and has retry support if for any reason the auth token expires or is invalided. The UI will not wait to render its own iframe until the /sync api endpoint (which returns the groups) is resolved --this speeds up load time.

Additional changes:
- Move tests into test/ folder for postmessage_json_rpc/server
- Create index-test.js for postmessage_json_rpc/server/index.js
- Remove postmessage_json_rpc/server/methods.js and migrate its content into postmessage_json_rpc/server/index.js
- add waitForElementToBeRemoved test helper