Remove `hide` prop from Spinner component
This was added as an affordance for writing tests to make it easier to
write a test that waits for a spinner to be removed. However this
approach requires an extra test-only prop for each component where we
want to check for this.

This commit instead uses a more generic approach of waiting for an
element to be removed:

```
await waitFor(() => {
  wrapper.update();
  return !wrapper.exists('ComponentName');
});
```

This is wrapped up into `spinnerHidden` and `spinnerVisible` utilities
since it would otherwise be repeated a few times.