Add MVP for launching VitalSource book viewer from LMS app
This commit adds a "VitalSource Book" option to the file picker app which
creates an assignment/activity using a hard-coded VitalSource book and chapter.
When that assignment is launched, the VitalSource book viewer is loaded
inside an iframe showing the specified book and chapter.

In future our LMS app will query the VitalSource API for a list of
available books and chapters within a selected book, for use as the text
for an assignment.

The initial VS integration adds two new configuration variables:
VITALSOURCE_LAUNCH_KEY and VITALSOURCE_LAUNCH_SECRET. These are
credentials used to sign the OAuth 1.0-authenticated LTI launch request
used to load the VitalSource book viewer. In future there will also be
an API key for authenticating access to the VS API.

Changes in detail:

 - Add a backend service in `lms/services/vitalsource.py`  to handle VitalSource
   LTI launches
 - Add a "Select book from VitalSource" button to the file picker app
   which submits the file picker form with an LTI launch URL for a VS
   book.
 - Add logic to the LTI launch views to handle launching an assignment
   configured to show a VS book
 - Add logic to the frontend in `BasicLtiLaunchApp` and
   `VitalSourceBookViewer` to load the VS book viewer inside an iframe.
 - Whitelist the `VITALSOURCE_LAUNCH_KEY` and
   `VITALSOURCE_LAUNCH_SECRET` keys in tox.ini. These keys are needed by
   the VitalSource launch request