Add "VitalSource Book" option to file picker behind feature flag
This commit adds a option to the file picker app which configures an
assignment that will use a hard-coded VitalSource book and chapter. This
assignment configuration is currently not handled by the LTI launch
code. This will follow in subsequent PRs.

This option is currently behind a `vitalsource` feature flag which is
enabled via cookies, to enable the VitalSource developers to turn it on
for themselves.

In future our LMS app will query the VitalSource API for a list of
available books and chapters within a selected book, for use as the text
for an assignment.