Support launching assignments using VitalSource books
This commit adds support for loading the VitalSource book viewer in an
LTI launch if the assignment URL contains `vitalsource_enabled=true` in
the query params, similar to how Canvas files assignments are
configured.

The VitalSource book viewer is loaded via the LTI 1.1 launch process.
When an assignment is configured to use VitalSource, the backend
generates form parameters for an LTI launch and signs them using a
key and secret specified by two new env vars: `VITALSOURCE_LAUNCH_KEY`
and `VITALSOURCE_LAUNCH_SECRET`. These form parameters are rendered into
the configuration for the LMS frontend.

When this configuration exists, the LMS frontend creates an empty iframe, renders
a form with these parameters into the iframe and auto-submits it. As a result of the
submission, the VitalSource book viewer app will load in the iframe at
the specified book and chapter and, if everything is correctly set up on
their end, Hypothesis should be loaded inside the app.

There is currently an issue that the VS app always tries to load the
production Hypothesis client, which won't work with the development LMS
app.

 - Whitelist the `VITALSOURCE_LAUNCH_KEY` and `VITALSOURCE_LAUNCH_SECRET` env vars in tox.ini.
 - Add a backend service in `lms/services/vitalsource.py` to handle VitalSource
   LTI launches
 - Add logic to the LTI launch views to handle launching an assignment
   configured to show a VS book
 - Add logic to the frontend in `BasicLtiLaunchApp` and
   `VitalSourceBookViewer` to load the VS book viewer inside an iframe.