Don't crash when dev key lacks scopes
Fixes https://github.com/hypothesis/lms/issues/1821

`CanvasAPIAuthorizeViews` is a little confusing and error-prone because:

* Some of the class's view methods (`authorize()` and
  `oauth2_redirect()`) are authenticated: `request.lti_user` is set for
  these views

* But the class's third view method (`oauth2_redirect_error()`) is not
  authenticated: `request.lti_user` is `None` for requests to this view.

The lack of authentication for the error view is unavoidable: the
request comes from Canvas (which redirects the popup window to us) and
is out of our control, and the request doesn't contain any kind of
parameter that we could use to authenticate the user.

The crash in https://github.com/hypothesis/lms/issues/1821 was caused
when a seemingly harmless refactor
(https://github.com/hypothesis/lms/pull/1812/) moved these two lines:

    self.request.find_service(name="ai_getter")
    self.request.find_service(name="course")

from the body of a view method into
`CanvasAPIAuthorizeViews.__init__()`. The `__init__()` is executed
whenever any of the class's views is requested, so this change meant
that all three views now depended on the `ai_getter` and `course`
services.

Both the `ai_getter` and `course` service factories depend on
`request.lti_user` being set. Since `request.lti_user` is `None` for the
`oauth2_redirect_error()` view, this meant that the error view now
crashes.

The `CanvasAPIAuthorizeViews` class has caused other confusions in the
past as well. For example many of its `@view_defaults` only apply to
some of its views and are overridden by others. And the inherting of
view defaults by exception views, though perfectly logical, has
surprised us in the past as well.

So I've decided to fix this by turning the three views into unrelated
view functions, and getting rid of the containing class.