UI improvements for dialog, spinner and error state of LMS app
The motivation for the extent of these changes comes from fixing a flickering issue when the error state jumps no error to an error between waiting for authorization of the popup window and receiving a second error from the api.

The BasicLtiLaunchApp is reworked to not reset error state on outgoing requests and instead resets error state value only after subsequent requests come back and we know there are no new errors.

The groups and contentUrl fetchers were previously fully atomic and took care of resetting the error state. They now return an error code which we can use to prevent flicker because the caller (authorizeAndFetchUrl) can now clear the error state rather than these two de-coupled requests.

Additionally some improvements were made to the spinner. These are mostly superficial but improve the UI. There is now an additional <Spinner> in the error dialog in addition to the full page spinner in BasicLtiLaunchApp. This has the advantage of showing the pending
requests while a previous error (if triggered) is still being displayed without rendering a huge spinner over the dialog.

One minor blind spot is if the initial requests fail and invoke an error. We don't keep track of which requests the loading states were triggered by. So for example, if the groups
errors first, then later the contentUrl resolves or errors, there will be a very short time the small spinner is rendered on the error dialog. This is unintended and cant not be solved unless we have a way of structuring the loading state with the errors that triggered them. This could be done either by a transaction or some queue or list. This would only add complexity and could be fixed later.

Additional fixes:

  - Spinner no longer operates by being removed from DOM. Now has `visible` prop. This gives it the advantage of having css animations. In some cases where we render a full background the spinner, it still gets removed from the DOM --that can be improved later to facilitate animations.
  - Allow styling to be set on the spinner's wrapper to adjust position from the caller component's css.
  - Refactor all instances of the spinner's css to use mixins.spinner()
  - Minor refactors on several tests that needed changes