Convert Canvas auth error dialog to a frontend mini-app
The Canvas API authorization error dialog re-used the visual style of
error dialogs elsewhere in our frontend applications but was server
rendered, re-using some of the same CSS and markup structure from client
components.

As the styling and markup of our base `Dialog` and `ErrorDisplay`
components has become more complex, we've encountered several scenarios
where the server and client rendered versions got out of sync or
functionality added to the client-rendered version was not present in
the server-rendered one.

This commit resolves the issue by converting the Canvas auth error
dialog into a mini-SPA which re-uses the existing frontend UI
components.