Automate `alembic stamp` in dev
In development environments automatically stamp the DB with the latest
version on startup if it isn't stamped already.

Fixes https://github.com/hypothesis/lms/issues/1751

Similar to what we already do in h:
https://github.com/hypothesis/h/blob/928fd87a4b2132716850733c99d8c016fda5356e/h/cli/commands/init.py#L33-L46

To test this:

1. Delete your DB:

       $ make services args='down'
       $ make services

2. Run `make dev` to recreate your DB:

       $ make dev

3. Query your current DB version:

       $ tox -qqe dev --run-command 'alembic -c conf/alembic.ini current'

4. Try to upgrade your DB to the latest version:

       $ tox -qqe dev --run-command 'alembic -c conf/alembic.ini upgrade head'

On master (3) would print out nothing because the DB isn't stamped and
(4) would try to run all the DB migrations from the first one and crash.

On this branch (3) will print out the latest version and (4) will run no
migrations since it knows you're already on the latest version.

fixup