Centralize frontend configuration parsing and documentation
Originally the LMS frontend and postmessage RPC server were entirely
separate and read configuration from different sources. For various
reasons the frontend app and RPC server were unified into a single
application that read the same configuration object, however the
configuration parsing was duplicated multiple times in the frontend.

This commit centralizes the knowledge of how to parse frontend
configuration from the page in a `readConfig` function in `config.js`.
Rather than having this function be called in multiple places, the
frontend's `index.js` calls this function once and then passes the
necessary configuration through to `startRpcServer` and the `Server`
constructor.

As part of this cleanup, JSDoc definitions of the structure of the
config object expected by the frontend have been added to `config.js`.
This provides a central reference that can be compared with the
corresponding `JSConfig` class on the backend.

 - Centralize frontend config parsing in `readConfig` function

 - Add JSDoc definitions of frontend config structure to `config.js`

 - Improve JSDoc types for `Server` class