Remove an incorrect comment
seanh flagged that this kind of information should be part of a docstring as it
contains information the caller should be aware of, except that it wasn't
completely correct anyway since there are non-LTI launch requests where
`request.lti_user` is set.

Since the exact requirements for instantiating this service may change in
future, I've removed it for now.