Add settings defaults and not-nullables
Make ApplicationInstance.settings and Course.settings default to empty
dicts and not-nullable at the Postgres level.

In practice there are no NULLs in the QA or prod DBs. Let's enforce that
at the DB level so that we don't need to worry about enforcing it (or
potentially failing to enforce it) at the Python level (either in Python
application code or in future DB migration scripts).