Convert `LMSGrader` props from prop-types to JSDoc
Improve the documentation and also use a shared definition of what a
"student" object looks like.

There is an open "TODO" in the code about changing a `username` property
to `userid` but I didn't make this change as the client's code still
references the `username` property [1] in one place without also
checking for `userid`.

[1] https://github.com/hypothesis/client/blob/ef82cf66d1a9616143d0bdb30c5a3da0cdd2ed4e/src/sidebar/store/modules/selection.js#L162