Copy all CI checks to GitHub Actions
This gets all the CI checks that we run on Jenkins (or that `make sure`
runs) working on GitHub Actions as well: backend and frontend
formatting, linting and unit, functional and BDD tests.

This runs two GitHub Actions jobs named backend and frontend in
parallel. Within the backend job, after installing everything, the
backend formatting, linting, and unit, functional and BDD tests all run
in parallel:

* In parallel:
  * backend:
    * In serial:
      - Start Postgres Docker container
      - Install Python 3.6.9
      - Update pip
      - Install tox
      - Checkout the lms git repo
      - Create the test databases
      - Install the frontend dependencies
      - Build the frontend assets
      - In parallel:
        - Run the backend code formatting
        - Run the backend linting
        - Run the backend unit tests
        - Run the backend functional tests
        - Run the backend BDD tests
  * frontend:
    * In serial:
      - Checkout the lms git repo
      - Install the frontend dependencies
      - Run the frontend code formatting
      - Run the frontend linting
      - Run the frontend tests

The .tox and node_modules are cached so that installing Python and
JavaScript dependencies doesn't need to be re-done for most builds.