Fix most remaining typecheck errors
This PR fixes the remaining errors when running `yarn typecheck`, except
for several related to a mixup of numbers and strings when formatting
grades that will be addressed separately.

One of the errors in `postmessage_json_rpc/server/index.js` indicates a
design issue that stems from an earlier version of the code where the
RPC server and LMS frontend were separate JS applications. As this PR is
focused on fixing typecheck errors I have just made a note of the issue and
added a check in case `getSidebarWindow` is called when the server has not
been started.