Restructure LMS frontend <-> Hypothesis client communication
The LMS frontend communicates with the Hypothesis client via an RPC `Server`
instance. Previously this was a singleton was started by a call to
`startRpcServer` in the LMS entry point. This function ostensibly returned a
`Server` object but it was actually augmented with some additional LMS-specific
methods/properties. Some UI components received a reference to this server
instance directly and others called functions (eg.  `getSidebarWindow`) that
accessed the singleton instance via the `server` global in
`postmessage_json_rpc/server/index.js`.

Apart from the inconsistency of calling methods on the server vs using
functions that depended on global state, this also meant that the details of
communication between the LMS frontend and Hypothesis client were not
encapsulated in one place but leaked into certain UI components (eg.
`LMSGrader`).

This commit restructures this code to something much closer to the way
the Hypothesis client is structured using _services_. Services are classes in
a `services/` dir which are instantiated on startup and then passed to the
components that need them via props. In the client we have a
dependency injection system that avoids the need to pass a service down
many layers, but we don't really have a need for that yet in the LMS
frontend because the component tree is quite shallow.

 - Add a `ClientRpc` class in `frontend_apps/services/client-rpc` which wraps
   the `Server` object with methods that provide a high level API for
   communicating with the Hypothesis client. An instance of this class
   is constructed by the LMS entry point when showing the
   `BasicLtiLaunchApp` app and passed down to the components that need
   it.

   This class provides better encapsulation of LMS frontend <-> client
   communication so that the `LMSGrader` class for example no longer
   deals with the details of the RPC call to change the focused user.

 - Simplify the file structure of the `postmessage_json_rpc` directory
   by putting everything in one directory rather than having separate
   `client` and `server` directories. This follows our conventions in
   frontend code to keep directory structure fairly shallow.