Convert remaining UI component docs to JSDoc
This surfaced an issue where several properties were missing from the
`StudentInfo` type listed in `frontend_apps/config.js`. These properties are
needed when calling the API to submit a grade update.

It also surfaced an issue where the `SubmitGradeForm` component redundantly took
a nullable `student` prop as well as a `disabled` prop.  Whether to disable the
form can be inferred from whether `student` is null or not (and _should_ be to
avoid a possible inconsistency).