Fix the Course test factory
Test factories for SQLAlchemy models should have
`FACTORY_CLASS=SQLAlchemyModelFactory`. This lets Factory Boy know that
it's a SQLAlchemy factory and means that it automatically adds it to the
test DB session when you do `factories.course()`,
`factories.course.create()` or `factories.course.create_batch()`.

If you want to make one _without_ adding it to the DB session (or
without even needing to have a DB session) use `build()` or
`build_batch()`.

See `factories.GradingInfo` for an existing example of this.