Remove some unnecessary comments
"Separator comments" aren't really part of our code style.

You can tell when something's a fixture because it has an
`@pytest.fixture` decorator on it. You can tell when something is a
helper method because it neither has the `@pytest.fixture` decorator nor
has a name beginning with `test_`. You can tell when something is an
assertion method (a special kind of test helper method) because it has a
name beginning with `assert_`.

It's a good idea to group these different kind of methods together but
we don't need separator comments for them, and we never use these kind
of comments anywhere else in our code.

The comments liable to become misleading when people modify things or
add new things in the wrong places.