Make `mockImportedComponents` work for components that don't use prop-types
We're about to remove prop-types from UI components, so change
`mockImportedComponents` not to rely on the presence of `propTypes`
property.

Unfortunately there isn't an obvious way to differentiate a component
from a function or class constructor with complete certainty. This
commit implements a hack which is fairly simple and can be adjusted as
necessary:

 - The function name starts with a capital letter
 - The function's source contains a `return <JSX expression>` statement.
   This second criterion is implemented with a Good Enough™ regex.