Pin all dev, test, lint, etc requirements
List all direct and indirect dependencies for all tox environments (dev,
test, lint, formatting, ...) with pinned versions in
`requirements-*.txt` files.

This is so that all of our tox environments (dev, lint, formatting,
tests, ...) will always contain the exact same versions of the exact
same packages on every developer's machine and on CI. When a new version
of some package gets released (e.g. PyLint) this will *not* cause our
dev and CI envs to change and potentially break. Instead, Dependabot
will send PRs to upgrade the necessary `requirements-*.txt` files and,
if there's any breakage, we will see that on the Dependabot PR and the
breakage won't get onto master.

This fixes an issue that new versions of PyLint, pytest, isort, etc etc
are often released and cause CI or dev environment commands to break.

1. Don't list any requirements directly in `tox.ini`.

   `tox.ini` now only refers to requirements files:

   ```ini
   deps =
       dev: -r requirements-dev.txt
       {format,checkformatting}: -r requirements-format.txt
       lint: -r requirements-lint.txt
       tests: -r requirements-tests.txt
       functests: -r requirements-functests.txt
       bddtests: -r requirements-bddtests.txt
       dockercompose: -r requirements-dockercompose.txt
       pipcompile: -r requirements-pipcompile.txt
   ```

2. List all requirements in `requirements-*.in` files.

   For every environment used in `tox.ini` there's now a corresponding
   `requirements-ENV.in` file. These files list direct dependencies only
   and don't pin versions.

   These `requirements-*.in` files are **only** used to compile the
   `requirements-*.txt` files (see below), they're not used or
   referenced by anything else. Dependabot doesn't use them. The
   `tox.ini` file doesn't use them.

3. Compile all `requirements-*.in` files to `requirements-*.txt` files.

   Compile each `requirements-ENV.in` file to a corresponding
   `requirements-ENV.txt` file that lists all direct and indirect
   dependencies pinned to exact versions.

   These `requirements-*.txt` files are the files that Dependabot pull
   requests will modify.

   They're also the files that `tox.ini` refers to.

   The command to compile a `requirements-ENV.in` file is:

   ```terminal
   tox -e pipcompile -- requirements-ENV.in
   ```