Prevent dialog size from changing when error details opened
By default the `Dialog` component's width is set based on the preferred
width of the dialog content, up to a default maximum of min(700px,
90vw). In dialogs that have an "Error details" section rendered by an
`ErrorDisplay`, this can cause the dialog's width to change when the
error details are expanded and collapsed. When the `ErrorDisplay` is
collapsed the preferred width depends only on the `<summary>`. When
open, it also depends upon the content, which by default is the laid-out
width of the error message before any line breaking happens.

This commit resolves the issue by setting `width` to override the
intrinsic width of the error message which is used when calculating the
width of the dialog but then also `min-width` to make the error message
fill the final width allocated to the `ErrorDisplay`. Hat tip to [1].

Several `max-width` properties were removed as they are no longer
required.

Fixes #1890

[1] https://stackoverflow.com/questions/24632208/