Render error messages as sentences
Add trailing periods to error messages if they do not already end with
punctuation.

The error messages come from a variety of sources so it is difficult to
ensure that they are always consistent about whether or not they end
with punctuation. Therefore we treat this as a formatting concern and
add it if necessary when rendering the error.

There was one place (in `LMSFilePicker`) where the `message` prop was
not a string. Change this to a string so that we don't have to deal with
the non-string case in `ErrorDisplay`. Also enable mocking of imported
components by default in `LMSFilePicker` tests to avoid implementation
changes in imported components breaking these tests.

Fixes #1891