Add UI to select books from VitalSource
Modify the "Select book from VitalSource" option in the file picker app
to show a book and chapter selection UI instead of submitting a
hardcoded book. The data shown in the book and chapter selection UI is
currently hardcoded data, derived from the VitalSource Products API [1],
for the two books which we know are available.

 - Add `BookPicker`, `BookList` and `ChapterList` components to support
   selecting a book from VitalSource

 - Modify `ContentSelector` component to show the `BookPicker` when the
   "Select book from VitalSource" option is selected, instead of
   submitting a hardcoded book ID and chapter.

 - Add dummy book and chapter data from the VitalSource API in
   `vitalsource-sample-data.js`. This will in future be replaced by
   calls from the frontend to the backend which will in turn make a
   request to the VitalSource Products API.

[1] https://developer.vitalsource.com/hc/en-us/articles/360010967153-GET-v4-products-vbid-Title-TOC-Metadata