Revert `LMSFilePicker` dialog focus changes
Revert the changes in the previous commit to focus buttons in the
`LMSFilePicker` dialog. This change didn't work because the `Select`
button is disabled when the dialog is initially displayed and the list
of files is being fetched, so trying to focus it had no effect [1].

The result was that focus remained on the "Select PDF file from Canvas"
button _behind_ the dialog instead of the dialog content.

[1] See https://html.spec.whatwg.org/multipage/semantics-other.html#concept-element-disabled