Ignore `initialFocus` if element is `null` or disabled
Ignore the `initialFocus` prop passed to `Dialog` and focus the root
element of the dialog if the `initialFocus` element is not rendered or
disabled at the point when the focusing logic runs.

Such a scenario may indicate a logic error in the parent of `Dialog` but
if that happens it is better to focus the dialog than cause an error, which
will most likely leave focus on whatever element launched the dialog.