Remove `aria-describedby` wrapper around dialog children
b04e102270ddd01eeac2b25d826a143d0f03e7c1 added a div with an ID that
wrapped the contents of a dialog, that was then referenced by `aria-describedby`
in order to cause screen readers to announce the text of the dialog when opened.

Unfortunately wrapping the children broke the layout of the Canvas file picker
when there were many files in the table - see #2003. In addition,
labeling the entire contents of the dialog as the description is not
always appropriate. Typically only static text (eg. the first paragraph)
should be labeled this way.

This commit reverts the addition of `aria-describedby` to fix the
problem with the Canvas file picker while we look for a better solution.