Re-introduce accessible descriptions for dialogs
This is a follow-up to https://github.com/hypothesis/lms/pull/2018 which
re-introduces accessible descriptions for dialogs using a different
approach. Instead of wrapping all of the dialog contents in an element
which is marked as the description, a layout effect instead picks out
the first `<p>` element, if any, and marks it as the description.

Compared to the previous approach this has some benefits:

 - It only identifies static text as a description, not non-descriptive
   elements like tables or images

 - It doesn't wrap the children, which avoids the issue that was fixed
   in #2018

There is also a downside in that it won't update if the dialog's content
changes after it is initially displayed. I believe this is an acceptible
limitation.

There are other approaches we could consider such as requiring `Dialog`
consumers to explicitly specify a separate `description` prop. That
would be more React-y, although it may duplicate the content. The
approach taken here has the benefit of just-working with existing
dialogs.