Disable "Authorize" button when re-fetching groups
After completing OAuth authorization when launching an assignment, the
"Authorize" button should be disabled while the LMS frontend waits for
both content URL and group fetching to complete. Previously the button
was disabled while the API request to fetch the content URL was pending,
but not if the content URL had been fetched and group fetching was still
in progress.

This commit resolves that issue by adding conditional logic to the
`fetchGroups` helper to control whether it updates the `fetchCount`
state that tracks the number of API requests in flight that block
display of the content. This logic is conditional because fetching
groups does not block content display on the initial launch but it does
after authorization, to avoid a flicker (as explained in #1826) if
authorization fails a second time.

Part of #1826